4.1.0 Added support for Maya 2020. Fixed an offset issue when more than one controller input is used.

+++ source/weightsServer/weightsServer.cpp
@@ -1188,7 +1188,7 @@ MStatus weightsServer::compute(const MPlug& plug, MDataBlock& data)
                                 influenceValue *= curveValue;
                             }
 
-                            value += influenceValue * scaleArray[j] * globalInfluenceScaleVal;
+                            value += influenceValue * (float)scaleArray[j] * (float)globalInfluenceScaleVal;
                             outputWeights.set(value, m);
                         }
                     }
